// src/interface/enum.ts
export enum HttpStatusMessage {
    CONTINUE = 'CONTINUE',
    SWITCHING_PROTOCOLS = 'SWITCHING_PROTOCOLS',
    PROCESSING = 'PROCESSING',
    OK = 'OK',
    CREATED = 'CREATED',
    ACCEPTED = 'ACCEPTED',
    NON_AUTHORITATIVE_INFORMATION = 'NON_AUTHORITATIVE_INFORMATION',
    NO_CONTENT = 'NO_CONTENT',
    RESET_CONTENT = 'RESET_CONTENT',
    PARTIAL_CONTENT = 'PARTIAL_CONTENT',
    AMBIGUOUS = 'AMBIGUOUS',
    MOVED_PERMANENTLY = 'MOVED_PERMANENTLY',
    FOUND = 'FOUND',
    SEE_OTHER = 'SEE_OTHER',
    NOT_MODIFIED = 'NOT_MODIFIED',
    TEMPORARY_REDIRECT = 'TEMPORARY_REDIRECT',
    PERMANENT_REDIRECT = 'PERMANENT_REDIRECT',
    BAD_REQUEST = 'BAD_REQUEST',
    UNAUTHORIZED = 'UNAUTHORIZED',
    PAYMENT_REQUIRED = 'PAYMENT_REQUIRED',
    FORBIDDEN = 'FORBIDDEN',
    NOT_FOUND = 'NOT_FOUND',
    METHOD_NOT_ALLOWED = 'METHOD_NOT_ALLOWED',
    NOT_ACCEPTABLE = 'NOT_ACCEPTABLE',
    PROXY_AUTHENTICATION_REQUIRED = 'PROXY_AUTHENTICATION_REQUIRED',
    REQUEST_TIMEOUT = 'REQUEST_TIMEOUT',
    CONFLICT = 'CONFLICT',
    GONE = 'GONE',
    LENGTH_REQUIRED = 'LENGTH_REQUIRED',
    PRECONDITION_FAILED = 'PRECONDITION_FAILED',
    PAYLOAD_TOO_LARGE = 'PAYLOAD_TOO_LARGE',
    URI_TOO_LONG = 'URI_TOO_LONG',
    UNSUPPORTED_MEDIA_TYPE = 'UNSUPPORTED_MEDIA_TYPE',
    REQUESTED_RANGE_NOT_SATISFIABLE = 'REQUESTED_RANGE_NOT_SATISFIABLE',
    EXPECTATION_FAILED = 'EXPECTATION_FAILED',
    I_AM_A_TEAPOT = 'I_AM_A_TEAPOT',
    MISDIRECTED = 'MISDIRECTED',
    UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
    FAILED_DEPENDENCY = 'FAILED_DEPENDENCY',
    PRECONDITION_REQUIRED = 'PRECONDITION_REQUIRED',
    TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
    BAD_GATEWAY = 'BAD_GATEWAY',
    SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
    GATEWAY_TIMEOUT = 'GATEWAY_TIMEOUT',
    HTTP_VERSION_NOT_SUPPORTED = 'HTTP_VERSION_NOT_SUPPORTED',
    ACCOUNT_DELETED = 'ACCOUNT_DELETED',
    ACCOUNT_INACTIVE = 'ACCOUNT_INACTIVE',
    ACCOUNT_LOCKED = 'ACCOUNT_LOCKED',
    DEFAULT_MESSAGE = 'DEFAULT_MESSAGE',
    INVALID_IP77_PASSWORD = 'INVALID_IP77_PASSWORD',
}
export enum USER_TYPE {
    BUYER = "buyer",
    SELLER = "seller"
}
export enum COLLECTION {
    USER = 'user',
    SESSION = 'session',
    PRODUCT = 'product',
    CHAT = 'chat',
    ADMIN = 'admin',
    ADMIN_SESSION = 'adminSession',
    CART = 'cart',
    ORDER = 'ORDER'
 }
export enum ExceptionMessage {
    DEFAULT_MESSAGE = 'DEFAULT_MESSAGE',
    EMAIL_ALREADY_EXIST = 'EMAIL_ALREADY_EXIST',
    VERIFICATION_FAILED = 'VERIFICATION_FAILED',
    EMAIL_NOT_EXISTS = 'EMAIL_NOT_EXISTS',
    LOGIN_FAILED = 'LOGIN_FAILED',
    SOMETHING_WENT_WRONG = 'SOMETHING_WENT_WRONG',
    USER_NOT_EXISTS = 'USER_NOT_EXISTS',
    USER_ALREADY_EXIST = 'USER_ALREADY_EXIST',
    AUTH_INVALID_TOKEN = 'AUTH_INVALID_TOKEN',
    AUTH_INVALID_STRATEGY = 'AUTH_INVALID_STRATEGY',
    OTP_EXPIRED = 'OTP_EXPIRED',
    SESSION_NOT_FOUND = 'SESSION_NOT_FOUND',
    INVALID_REQUEST = 'INVALID_REQUEST',
    INCORRECT_OTP = 'INCORRECT_OTP',
    INVALID_APIKEY = 'INVALID_APIKEY',
    INVALID_OTP = 'INVALID_OTP',
    INVALID_PASSWORD = 'INVALID_PASSWORD',
    UNAUTHORIZED = 'UNAUTHORIZED',
    MOVED_PERMANENTLY = 'MOVED_PERMANENTLY',
    USER_BLOCKED = 'USER_BLOCKED',
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    TOKENIZATION_ERROR = 'TOKENIZATION_ERROR',
    LINK_EXPIRE = 'Link Expired',
    CONNECTION_NOT_FOUND = 'CONNECTION_NOT_FOUND',
    FILE_ERROR = 'FILE_NOT_FOUND',
    PRODUCT_NOT_EXSITS = 'PRODUCT_NOT_EXSITS',
    NOT_VALID_PRODUCT_ID = 'NOT_VALID_PRODUCT_ID',
    TOKEN_NOT_FOUND = "TOKEN_NOT_FOUND",
    ADMIN_NOT_FOUND = "ADMIN_NOT_FOUND",
    CART_DOES_NOT_EXSIST = "CART_DOES_NOT_EXSIST",
    ADDRESSS_NOT_FOUND = 'ADDRESSS_NOT_FOUND',
    INSERT_VALID_DATA = 'INSERT_VALID_DATA',
    ORDER_DATA_NOT_FOUND = 'ORDER_DATA_NOT_FOUND',
    ORDER_ALREADY_DELIVERED = 'ORDER_ALREADY_DELIVERED',
    USER_DATA_FOUND = 'USER_DATA_FOUND',
    ORDER_NOT_DELIVERED_YET =  'ORDER_NOT_DELIVERED_YET',
    REVIEW_NOT_FOUND = 'REVIEW_NOT_FOUND',
    ISSUE_IN_SUBSCRIBING = 'ISSUE_IN_SUBSCRIBING',
    FAILED_TO_LOGOUT = 'FAILED_TO_LOGOUT',
    FAILED_TO_VERIFY = 'FAILED_TO_VERIFY',
    FAILED_TO_CHANGE_PASSWORD = 'FAILED_TO_CHANGE_PASSWORD',
    MAIL_NOT_SENT = 'MAIL_NOT_SENT',
    PROBLEM_IN_ADDING_ADDRESS = 'PROBLEM_IN_ADDING_ADDRESS',
    FAILED_TO_ADD_ADDRESS =  'FAILED_TO_ADD_ADDRESS',
    INVALID_REACTION = 'INVALID_REACTION',
    MESSAGE_NOT_FOUND = 'MESSAGE_NOT_FOUND',
    ISSUE_IN_UNSUBSCRIBING = 'ISSUE_IN_UNSUBSCRIBING'
    
   
}
export enum SuccessMessage {
    PASSWORD_CHANGED = 'Password Change Successfully',
    Mail_SEND = 'Mail Send Successfully',
    USER_REGISTRATION_MAIL = 'OTP Send to entered email address',
    USER_REGISTRATION = 'Register successfully',
    PRODUCT_EDITED = 'PRODUCT_EDITED',
    PRODUCT_DELETED = 'PRODUCT_DELETED',
    PRODUCT_CREATED = 'PRODUCT_CREATED',
    ALL_PRODUCTS = 'ALL_PRODUCTS',
    PRODUCT_ADDED = 'PRODUCT_ADDED',
    CART_FOUND = 'CART_FOUND',
    CART_REMOVED = 'CART_REMOVED',
    ORDER_PLACED  =  'ORDER_PLACED',
    ORDER_CANCELED = 'ORDER_CANCELED',
    ORDER_FOUND = 'ORDER_FOUND',
    ALL_USER_DATA_FOUND = 'ALL_USER_DATA_FOUND',
    USER_ACCOUNT_REMOVED = 'USER_ACCOUNT_REMOVED',
    REVIEW_ADDED = 'REVIEW_ADDED',
    MESSAGE_PUBLISHED = 'MESSAGE_PUBLISHED',
    SUBSCRIBED_SUCCESSFULLY = 'SUBSCRIBED_SUCCESSFULLY',
    ALL_MESSAGES = 'ALL_MESSAGES',
    SUCCESSFULLY_LOGEDOUT = 'SUCCESSFULLY_LOGEDOUT',
    ADDRESS_ADDED  = 'ADDRESS_ADDED',
    ALL_CHAT_DATA_FOUND ='ALL_CHAT_DATA_FOUND',
    GROUP_MEMBERS_FOUND = 'GROUP_MEMBERS_FOUND'         
}

export enum OrderStatus {
    Active = 1,
    InActive = 2,
}

export enum WellcomeMsg {
    WELLCOME_TO_ADMIN_SERVICE = "WELLCOME_TO_ADMIN_SERVICE",
    WELLCOME_TO_USER_SERVICE = "WELLCOME_TO_USER_SERVICE"
}
export enum RedisExpirydata{
    ADMIN_LOGIN_DATA= 600,
    ADMIN_LOGIN_OTP = 600,
    USER_SINGIN_REDIS = 900,
}
export enum OTP {
    USER_OTP = 6,
    ADMIN_OTP = 4,
}